API: Add ability to project over subregions
Now you can get coverage or project over regions that are not
centered on the origin. This will be usefull to have in the
future for reducing computational workload by only computing over
a smaller region of interest or in order to break the workload
among many compute nodes by slicing the data.