API: Change parameters of and document scan.py functions
Documented building block functions for creating scanning
trajectories in scan.py. Removed the sampling frequency
parameter from these functions. Instead, the user supplies
an array of time values. This is better because it allows
irregular sampling along time dimension.

Changed the function of staircase and sawtooth because the period
and amplitude of these functions did not match what I expected.