BUG: euclidian_distance is numerically unstable
Using the law of cosines to simplify the euclidian_distance leads
to sometimes calculating the sqrt of very small negative numbers.
I think this is due to rounding error because the numbers are
on the order of `-1e-9`. I added an `abs` function to prevent
imaginary roots.

I also added an approximate euclidian distance function which
calculates the distance by adding the arclength of the move
to the distance in the hv plane. It's not strictly greater
than the actual distance function.