ENH: trim_coords using binary search
When trimming line / grid intersections that are ouside the grid,
use a binary search to find edges of region inside the grid instead
of exhaustively checking every point.

This works because we know that the points along the line are
ordered. i.e. if points 3 and 5 are inside the grid, then we
know that point 4 is also in the grid.