BUG: Indices out of bounds due to numerical instability
indx and indy in calc_index would be out of bounds because of
rounding errors in the expression midx - oxmin. In order to keep
this result in bounds, we can divide by 2 + (small number) when
calculating the midpoint to make it slightly smaller such that
it will remain above zero if it is supposed to be zero, but
it will move away from the upper boundary of the grid.