REF: Refactor tomo.c backend to match xdesign implementation
Moved all interection calculations to separate file and header
siddon.h and siddon.c in order to make tomo.c shorter and make
reusing code for all of the methods that forward and back
projection easier. This also enables switching out the
intersection calculation in less time because that part of modular
and separate.