BUG: Fix improper indexing of z-slices when anisotropy is False
It could occur that anisotropy param of coverage functions would
be false. This would be passed to the coverage function as an
ot parameter of 0. The result was improper slicing along the z
direction as every slice would start at the beginning of the
array. Also an additional test was added to check slicing
correctness.