API: Remove theta from required arguments in tike.ptycho
Although this makes the internal workings more awkward because
now data.shape[0] != h.size. It means that for tike.tike there
is only one theta parameter, and tike.ptycho doesnt need to know
what angles each of the measurements are at. This is actually not
good for data acquisition flexibility, but it's how people are
currently collecting data...