API: Rename foo_min to foo_corner
This commit replaces all uses of foo_min with foo_corner;
foo is a placeholder for obj, data, probe, etc.

@dgursoy and I agree that foo_min is ambigious because it
could mean the same as min(foo). foo_corner is clearer
because it could only mean the corner of foo; the next
logical question of "which corner?" can be gleaned from the
various docstrings.

It is worth it to do this now because there are only 3
people whose code will be disrupted by this change.