STY: Change how functions defs are wrapped when too long
This new method of wrapping is better than the way suggested
in PEP8 because in a text editor, you can fold the function
implementation while still viewing the arguments. This is
good because python doesn't have header files, so you have
to look at the code in order see its signature.