BUG: Prevent zero division in np.angle replacement
We previously replaced numpy angle with a cheaper alternative
but because of zero division it did not behave the same as
np.angle. This patch makes the trig identity behave the same way
as numpy angle when the complex number is zero. nan and inf were
not considered in this patch.