REF: Combine ptycho regularizer params: h and lamda
Replace the two parameters h and lamda for tike.ptycho.grad
with one param: reg, which is h - lamda/rho. We are doing this
to reduce the regularization parameters to one parameter. They
must remain separate params in tike.tike because of the lamda is
used to switch to the linear tomography problem while h is not.