REF: Make ptycho solvers compatible with new pytchocg API
Because:
- Modularizing the backend and allows code reuse
- Modularizing the backend makes future upgrades easier

This reogranization makes the ptychographic operators
separate from the solver algorithm implementations.
Basically, what I have done is put the algorithms and
ptychographic operators such as the forward and reverse
operator into classes. Each algorithm is a class that
inherits from the operator class.

I also mostly removed the fast_pad usage because I
discovered that the fft libraries will do that for you.