API: Replace v, h, with scan parameter and reshape arrays
Because:
- Compatibility with ptychocg and experimental data
- Reconstruct multiple ptycho views at once sharing a context

This commit replaces the two separate parameters v, h with a single scan parameter
that is a combined array of shape (ntheta, nscan, 2). This is more likely to be a
data shape that experimentalists use. Also, an additional dimension is added to psi,
data, farfield, scan, etc. this dimension respresents the view and corresponds to
ntheta.