API: Restructure core operators
Collect all of the operators into a central module and break the
ptychography operator into the composition of two operators:
convolution and propagation. This makes the framework more modular.
Operators now all inherit from a Operator base class and the
interface for each operator type is now combined with the default
implementation.