API: Incorporate adjoint_probe into Convolution operator
It is more conventient to have an adjoint operator with
respect to the probe than it is to manually create the
adjoint probe operator from the fwd operator every time.
A survey of the code shows that the fwd operator alone
(without immediately multiplying by the probe) only occurs
once.