NEW: Add drift correction to position correction
In order to prevent the reconstruction drifting because all the
points move in the same direction, we can enforce that the center
of mass of the scan positions remains the same. This allows them
to move freely relative to one another, but forces the blob to
stay in its initial position.