NEW: autoinitialize the object if not given
When the user provides no initial value for the object in
ptycho.reconstruct, the object is now automatically initialized
with a complex ones array such that field of view is padded by
one eight of the width of the field of view on each side. The
scan positions are also update to be centered in this field of
view.