REF: Change initial step sizes for line search
From my observations of two datasets. It seems that a better
initial step size is approx. 1e-5 for the object updates and
1 for the probe updates. This tweak should decrease the amount
of time spent line searching. There is no logic behind these updates,
they just seem to work. I observed an order of magnitude decrease
for the time spend on one iteration.