REF: Skip repeated line searches for mini-batches
With the new mini-batch feature, it is not necessary to do a new
line search with each mini-batch because the step size for each
mini-batch appears to be on the same order of magnitude. Thus, we
can do the line search for the first mini-batch of each epoch and
reuse the step size afterards.

conjugate gradient is also disabled by default because it appears
that mini-batches are much better at driving down the cost
function than it is using full-batch + conjugate gradients.