BUG: Invalid algorithms not caught when matching submodule name
The check for invalid algorithm names fails when the algorithm
name matches any object in the solvers namespace. The solution
for this is the check against the contents of an explicitly set
__all__ parameter which will only ever contain the names of
valid solver functions.

Previously dir(solvers) was used in order to automate the
generation of this list. However, it seems that this is unsafe.