BUG: Fix background mismatches of sub-images reconstruted by different processes.
The first issue is the probe randomization in tests/test_ptycho.py.
It made different processes have different probes. A MPI Bcast is
used to solve this issue.

The second issue is _rescale_obj_probe in ptycho.py. The probe rescaling
uses data and scan but previously was not sync and communicated. Hence
it also made different processes have different probes.