API: Allow out of bounds patches, but skip them
Now that position correction is moving scanning
positions around, it is likely that some scanning
positions will move to overlap the edge of the field
of view or go completely out of bounds. Instead of
triggering an assertion error in the CUDA kernel,
we now just skip patch extraction for the pixels
which are out of bounds. This allows partial patches
and just uses whatever value was in the initialized
patch array.