API: Automatically collect times and costs into single list
The "cost" parameter has been renamed "costs" to be consistent
with "times".

Previously the objective value at the end of each iteration was
collected in a list then converted into a numpy array and appended.
Now these lists (for iteration time and objective value) are
assumed to be lists and directly appended each iteration.

Also fixed a bug where the cgrad method would sometimes return
a cost that was not localized to host memory.