TST: Remove tests that rely on undocumented behavior in numpy
This smoke test for the clustering methods relies on argpartition
returning the same result every time. However, the behavior of
argparition changed with NumPy 2.0. Fortunately(?) the behavior
change is not a bug because argparition never promised a stable
paritioning, it only promises that elements will correctly be
divided before and after the pivot. However, these test must
be removed because they are recall tests which assume that every
argparition implementation returns the exact same result.